move custom gates to seperate module, moved importing simulator into a function call
