+++ .gitignore
@@ -127,6 +127,6 @@ test/python/test_save.json
 
 test/python/*.tex
 
-external/cpp-simulator/local_qiskit_simulator
+src/cpp-simulator/local_qiskit_simulator
 
-external/cpp-simulator/src/main.o
+src/cpp-simulator/src/main.o
